Engine API-independent shader binding
