+++ BRenderer/Renderer/Descriptors.h
@@ -1,75 +1,32 @@
 #ifndef BRR_DESCRIPTORS_H
 #define BRR_DESCRIPTORS_H
-#include <Renderer/Vulkan/VulkanInc.h>
+#include <Renderer/Vulkan/VulkanRenderDevice.h>
+#include <Renderer/RenderEnums.h>
 #include <Core/LogSystem.h>
 
 namespace brr::render
 {
-    //--------------------------------------------//
-    //-----------  DescriptorAllocator  ----------//
-    //--------------------------------------------//
-    class DescriptorAllocator
-    {
-    public:
-        DescriptorAllocator() = default;
-        explicit DescriptorAllocator(vk::Device device);
-
-        ~DescriptorAllocator();
-
-        void ResetPools();
-        bool Allocate(const std::vector<vk::DescriptorSetLayout>& layouts, std::vector<vk::DescriptorSet>& out_set);
-
-        void Cleanup();
-
-        [[nodiscard]] vk::Device GetDevice() const { return m_device; }
-
-        [[nodiscard]] bool IsValid() const { return m_device; }
-
-        struct PoolSizes
-        {
-            std::vector<std::pair<vk::DescriptorType, float>> sizes_multiplier
-            {
-                { vk::DescriptorType::eSampler, 0.5f },
-                { vk::DescriptorType::eCombinedImageSampler, 4.f },
-                { vk::DescriptorType::eSampledImage, 4.f },
-                { vk::DescriptorType::eStorageImage, 1.f },
-                { vk::DescriptorType::eUniformTexelBuffer, 1.f },
-                { vk::DescriptorType::eStorageTexelBuffer, 1.f },
-                { vk::DescriptorType::eUniformBuffer, 2.f },
-                { vk::DescriptorType::eStorageBuffer, 2.f },
-                { vk::DescriptorType::eUniformBufferDynamic, 1.f },
-                { vk::DescriptorType::eStorageBufferDynamic, 1.f },
-                { vk::DescriptorType::eInputAttachment, 0.5f },
-            };
-        };
-
-    private:
-
-        vk::DescriptorPool GrabPool();
-
-        vk::Device m_device {};
-
-        vk::DescriptorPool current_pool_ {};
-        PoolSizes descriptor_pool_sizes_;
-
-        std::vector<vk::DescriptorPool> used_pools_ {};
-        std::vector<vk::DescriptorPool> free_pools_ {};
-    };
+    class VulkanRenderDevice;
 
     //--------------------------------------------//
     //--------  DescriptorLayoutBindings  --------//
     //--------------------------------------------//
+    struct DescriptorLayoutBinding
+    {
+        uint32_t binding;
+        DescriptorType descriptor_type;
+        ShaderStageFlag shader_stage_flag;
+    };
+
     struct DescriptorLayoutBindings
     {
         DescriptorLayoutBindings() = default;
 
         bool operator==(const DescriptorLayoutBindings& other) const;
 
-        vk::DescriptorSetLayoutBinding& operator [] (uint32_t binding) { return m_bindings[binding]; }
+        DescriptorLayoutBinding& operator [] (uint32_t binding) { return m_bindings[binding]; }
 
-        size_t Hash() const;
-
-        std::vector<vk::DescriptorSetLayoutBinding> m_bindings {};
+        std::vector<DescriptorLayoutBinding> m_bindings {};
     };
 
     //--------------------------------------------//
@@ -77,39 +34,8 @@ namespace brr::render
     //--------------------------------------------//
     struct DescriptorLayout
     {
-        vk::DescriptorSetLayout  m_descriptor_set_layout;
+        DescriptorLayoutHandle m_layout_handle;
         DescriptorLayoutBindings m_bindings;
-    };
-
-    //--------------------------------------------//
-    //----------  DescriptorLayoutCache  ---------//
-    //--------------------------------------------//
-    class DescriptorLayoutCache
-    {
-    public:
-        DescriptorLayoutCache() = default;
-        explicit DescriptorLayoutCache(vk::Device device);
-
-        ~DescriptorLayoutCache();
-
-        vk::DescriptorSetLayout CreateDescriptorLayout(DescriptorLayoutBindings layout_bindings);
-
-        void Cleanup();
-
-        [[nodiscard]] bool IsValid() const { return m_device; }
-
-    private:
-
-        struct DescriptorLayoutHash
-        {
-            size_t operator()(const DescriptorLayoutBindings& info) const
-            {
-                return info.Hash();
-            }
-        };
-
-        std::unordered_map<DescriptorLayoutBindings, vk::DescriptorSetLayout, DescriptorLayoutHash> m_layoutCache;
-        vk::Device m_device {};
     };
 
     //--------------------------------------------//
@@ -120,11 +46,11 @@ namespace brr::render
 	public:
         DescriptorLayoutBuilder() = delete;
 
-        static DescriptorLayoutBuilder MakeDescriptorLayoutBuilder(DescriptorLayoutCache* layoutCache);
+        static DescriptorLayoutBuilder MakeDescriptorLayoutBuilder(VulkanRenderDevice* render_device);
 
         DescriptorLayoutBuilder& SetBinding(uint32_t binding,
-								            vk::DescriptorType type,
-								            vk::ShaderStageFlags stageFlags);
+								            DescriptorType type,
+								            ShaderStageFlag stageFlags);
 
         [[nodiscard]] DescriptorLayout BuildDescriptorLayout();
 
@@ -132,13 +58,13 @@ namespace brr::render
         template <uint32_t N_Sets>
         friend class DescriptorSetBuilder;
 
-        DescriptorLayoutBuilder(DescriptorLayoutCache* layout_cache)
-        : m_layoutCache(layout_cache)
+        DescriptorLayoutBuilder(VulkanRenderDevice* render_device)
+        : m_render_device(render_device)
         {}
 
         DescriptorLayoutBindings m_descriptor_layout_bindings {};
 
-        DescriptorLayoutCache* m_layoutCache = nullptr;
+        VulkanRenderDevice* m_render_device = nullptr;
     };
 
 
@@ -150,37 +76,42 @@ namespace brr::render
     public:
         DescriptorSetBuilder() = delete;
 
-        static DescriptorSetBuilder MakeDescriptorSetBuilder(const DescriptorLayout& layout, DescriptorAllocator* descriptor_allocator);
-        static DescriptorSetBuilder MakeDescriptorSetBuilder(DescriptorLayoutBuilder& layout_builder, DescriptorAllocator* descriptor_allocator);
+        static DescriptorSetBuilder MakeDescriptorSetBuilder(const DescriptorLayout& layout, VulkanRenderDevice* render_device);
+        static DescriptorSetBuilder MakeDescriptorSetBuilder(DescriptorLayoutBuilder& layout_builder, VulkanRenderDevice* render_device);
 
-        DescriptorSetBuilder& BindBuffer(uint32_t binding, std::array<vk::DescriptorBufferInfo, N_Sets> bufferInfo);
+        DescriptorSetBuilder& BindBuffer(uint32_t binding, std::array<BufferHandle, N_Sets> buffers_info);
 
-        DescriptorSetBuilder& BindImage (uint32_t binding, std::array<vk::DescriptorImageInfo,  N_Sets> imageInfo);
+        DescriptorSetBuilder& BindImage (uint32_t binding, std::array<Texture2DHandle,  N_Sets> images_info);
 
         
-        bool BuildDescriptorSet(std::array<vk::DescriptorSet, N_Sets>& sets, vk::DescriptorSetLayout& layout);
-        bool BuildDescriptorSet(std::array<vk::DescriptorSet, N_Sets>& sets);
+        bool BuildDescriptorSet(std::array<DescriptorSetHandle, N_Sets>& sets, DescriptorLayoutHandle& layout);
+        bool BuildDescriptorSet(std::array<DescriptorSetHandle, N_Sets>& sets);
 
     private:
-        DescriptorSetBuilder(const DescriptorLayout& layout, DescriptorAllocator* descriptor_allocator)
-        : m_descriptor_layout(layout), m_descriptorAlloc(descriptor_allocator)
+        DescriptorSetBuilder(const DescriptorLayout& layout, VulkanRenderDevice* render_device)
+        : m_descriptor_layout(layout), m_render_device(render_device)
         {}
 
-        std::array<std::vector<vk::WriteDescriptorSet>, N_Sets> m_descriptorWrites;
+        struct ShaderSetBinding
+        {
+            BufferHandle buffer_handle = null_handle;
+            Texture2DHandle texture_handle = null_handle;
+        };
 
+        std::array<std::vector<ShaderSetBinding>, N_Sets> m_descriptor_writes;
 
         DescriptorLayout            m_descriptor_layout;
-        DescriptorAllocator*        m_descriptorAlloc;
+        VulkanRenderDevice*         m_render_device;
     };
 
     template <uint32_t N_Sets>
     DescriptorSetBuilder<N_Sets>
     DescriptorSetBuilder<N_Sets>::MakeDescriptorSetBuilder(
         const DescriptorLayout& layout,
-        DescriptorAllocator* descriptor_allocator
+        VulkanRenderDevice* render_device
     )
     {
-        DescriptorSetBuilder set_builder {layout, descriptor_allocator};
+        DescriptorSetBuilder set_builder {layout, render_device};
 
         return set_builder;
     }
@@ -188,10 +119,10 @@ namespace brr::render
     template <uint32_t N_Sets>
     DescriptorSetBuilder<N_Sets> DescriptorSetBuilder<N_Sets>::MakeDescriptorSetBuilder(
         DescriptorLayoutBuilder& layout_builder, 
-        DescriptorAllocator* descriptor_allocator
+        VulkanRenderDevice* render_device
     )
     {
-        DescriptorSetBuilder set_builder { layout_builder.BuildDescriptorLayout(), descriptor_allocator };
+        DescriptorSetBuilder set_builder { layout_builder.BuildDescriptorLayout(), render_device };
 
         return set_builder;
     }
@@ -199,27 +130,31 @@ namespace brr::render
     template <uint32_t N_Sets>
     DescriptorSetBuilder<N_Sets>& DescriptorSetBuilder<N_Sets>::BindBuffer(
         uint32_t binding,
-        std::array<vk::DescriptorBufferInfo, N_Sets> bufferInfo
+        std::array<BufferHandle, N_Sets> bufferInfo
     )
     {
-        if (!m_descriptorAlloc)
+        if (!m_render_device)
         {
             BRR_LogError("Can't bind buffer with non-initialized DescriptorSetBuilder");
             return *this;
         }
 
-        vk::DescriptorSetLayoutBinding descriptor_binding = m_descriptor_layout.m_bindings[binding];
+        //DescriptorLayoutBinding descriptor_binding = m_descriptor_layout.m_bindings[binding];
+        //vk::DescriptorType vk_descriptor_type = VkHelpers::VkDescriptorTypeFromDescriptorType(descriptor_binding.descriptor_type);
+        
 
         for (uint32_t info_idx = 0; info_idx < N_Sets; info_idx++)
         {
-            vk::WriteDescriptorSet descriptor_write;
+            /*vk::WriteDescriptorSet descriptor_write;
             descriptor_write
                 .setPBufferInfo(&bufferInfo[info_idx])
                 .setDstBinding(binding)
-                .setDescriptorType(descriptor_binding.descriptorType)
-                .setDescriptorCount(1);
+                .setDescriptorType(vk_descriptor_type)
+                .setDescriptorCount(1);*/
 
-            m_descriptorWrites[info_idx].push_back(descriptor_write);
+            ShaderSetBinding descriptor_write;
+            descriptor_write.buffer_handle = bufferInfo[info_idx];
+            m_descriptor_writes[info_idx].push_back(descriptor_write);
         }
 
         return *this;
@@ -228,62 +163,74 @@ namespace brr::render
     template <uint32_t N_Sets>
     DescriptorSetBuilder<N_Sets>& DescriptorSetBuilder<N_Sets>::BindImage(
         uint32_t binding,
-        std::array<vk::DescriptorImageInfo, N_Sets> imageInfo
+        std::array<Texture2DHandle, N_Sets> imageInfo
     )
     {
-        if (!m_descriptorAlloc)
+        if (!m_render_device)
         {
             BRR_LogError("Can't bind image with non-initialized DescriptorSetBuilder");
             return *this;
         }
 
-        vk::DescriptorSetLayoutBinding descriptor_binding = m_descriptor_layout.m_bindings[binding];
+        //DescriptorLayoutBinding descriptor_binding = m_descriptor_layout.m_bindings[binding];
+        //vk::DescriptorType vk_descriptor_type = VkHelpers::VkDescriptorTypeFromDescriptorType(descriptor_binding.descriptor_type);
 
         for (uint32_t info_idx = 0; info_idx < N_Sets; info_idx++)
         {
-            vk::WriteDescriptorSet descriptor_write;
+            /*vk::WriteDescriptorSet descriptor_write;
             descriptor_write
                 .setPImageInfo(&imageInfo[info_idx])
                 .setDstBinding(binding)
-                .setDescriptorType(descriptor_binding.descriptorType)
-                .setDescriptorCount(1);
+                .setDescriptorType(vk_descriptor_type)
+                .setDescriptorCount(1);*/
 
-            m_descriptorWrites[info_idx].push_back(descriptor_write);
+            ShaderSetBinding descriptor_write;
+            descriptor_write.texture_handle = imageInfo[info_idx];
+            m_descriptor_writes[info_idx].push_back(descriptor_write);
         }
 
         return *this;
     }
 
     template <uint32_t N_Sets>
-    bool DescriptorSetBuilder<N_Sets>::BuildDescriptorSet(std::array<vk::DescriptorSet, N_Sets>& sets,
-                                                          vk::DescriptorSetLayout&               layout)
+    bool DescriptorSetBuilder<N_Sets>::BuildDescriptorSet(std::array<DescriptorSetHandle, N_Sets>& sets,
+                                                          DescriptorLayoutHandle&               layout)
     {
-        layout = m_descriptor_layout.m_descriptor_set_layout;
-        std::vector<vk::DescriptorSetLayout> layouts (N_Sets, layout);
+        layout = m_descriptor_layout.m_layout_handle;
+
+        std::array<std::vector<VKRD::DescriptorSetBinding>, N_Sets> descriptor_set_bindings;
+        for (uint32_t desc_idx = 0; desc_idx < N_Sets; desc_idx++)
+        {
+            auto& desc_write_vec = descriptor_set_bindings[desc_idx];
+            for (uint32_t binding_idx = 0; binding_idx < m_descriptor_writes[desc_idx].size(); binding_idx++)
+            {
+                VKRD::DescriptorSetBinding binding;
+                binding.descriptor_type = m_descriptor_layout.m_bindings[binding_idx].descriptor_type;
+                if (m_descriptor_writes[desc_idx][binding_idx].buffer_handle != null_handle)
+                {
+                    binding.buffer_handle = m_descriptor_writes[desc_idx][binding_idx].buffer_handle;
+                }
+                if (m_descriptor_writes[desc_idx][binding_idx].texture_handle != null_handle)
+                {
+                    binding.texture_handle = m_descriptor_writes[desc_idx][binding_idx].texture_handle;
+                }
+                desc_write_vec.push_back(binding);
+            }
+        }
 
         // Allocate descriptor
-        std::vector<vk::DescriptorSet> result_sets;
-        bool success = m_descriptorAlloc->Allocate(layouts, result_sets);
-        memcpy(sets.data(), result_sets.data(), N_Sets * sizeof(vk::DescriptorSet));
-        if (!success) { return false; };
-
-        for (uint32_t write_idx = 0; write_idx < N_Sets; write_idx++)
-        {
-            for (vk::WriteDescriptorSet& write : m_descriptorWrites[write_idx])
-            {
-                write.setDstSet(sets[write_idx]);
-            }
-
-            m_descriptorAlloc->GetDevice().updateDescriptorSets(m_descriptorWrites[write_idx], {});
-        }
+        std::vector<DescriptorSetHandle> result_sets = m_render_device->AllocateDescriptorSet(layout, N_Sets, descriptor_set_bindings);
+        bool                           success     = !result_sets.empty();
+        if (!success) { return false; }
+        memcpy(sets.data(), result_sets.data(), N_Sets * sizeof(DescriptorSetHandle));
 
         return true;
     }
 
     template <uint32_t N_Sets>
-    bool DescriptorSetBuilder<N_Sets>::BuildDescriptorSet(std::array<vk::DescriptorSet, N_Sets>& sets)
+    bool DescriptorSetBuilder<N_Sets>::BuildDescriptorSet(std::array<DescriptorSetHandle, N_Sets>& sets)
     {
-        vk::DescriptorSetLayout layout;
+        DescriptorLayoutHandle layout;
         return BuildDescriptorSet(sets, layout);
     }
 }
