+++ BRenderer/Renderer/Vulkan/VkInitializerHelper.cpp
@@ -306,4 +306,457 @@ namespace brr::render::VkHelpers
 
 		throw std::runtime_error("Failed to find valid memory type.");
     }
+
+    vk::DescriptorType VkDescriptorTypeFromDescriptorType(DescriptorType descriptor_type)
+    {
+		switch(descriptor_type)
+		{
+		case DescriptorType::UniformBuffer:
+            return vk::DescriptorType::eUniformBuffer;
+		case DescriptorType::StorageBuffer:
+            return vk::DescriptorType::eStorageBuffer;
+		case DescriptorType::CombinedImageSampler:
+            return vk::DescriptorType::eCombinedImageSampler;
+        case DescriptorType::SampledImage:
+            return vk::DescriptorType::eSampledImage;
+        case DescriptorType::StorageImage:
+            return vk::DescriptorType::eStorageImage;
+        case DescriptorType::Sampler:
+            return vk::DescriptorType::eSampler;
+        }
+		return vk::DescriptorType::eUniformBuffer;
+    }
+
+    vk::ShaderStageFlags VkShaderStageFlagFromShaderStageFlag(ShaderStageFlag shader_stage_flag)
+    {
+        vk::ShaderStageFlags shader_stage;
+		if  ((shader_stage_flag & ShaderStageFlag::VertexShader) != 0)
+		{
+			shader_stage |= vk::ShaderStageFlagBits::eVertex;
+		}
+		if  ((shader_stage_flag & ShaderStageFlag::FragmentShader) != 0)
+		{
+			shader_stage |= vk::ShaderStageFlagBits::eFragment;
+		}
+		if  ((shader_stage_flag & ShaderStageFlag::ComputeShader) != 0)
+		{
+			shader_stage |= vk::ShaderStageFlagBits::eCompute;
+		}
+		return shader_stage;
+    }
+
+    vk::Format VkFormatFromDeviceDataFormat(DataFormat data_format)
+    {
+		switch (data_format)
+        {
+        case DataFormat::R8_UNorm:
+            return vk::Format::eR8Unorm;
+        case DataFormat::R8_SNorm:
+            return vk::Format::eR8Snorm;
+        case DataFormat::R8_UScaled:
+            return vk::Format::eR8Uscaled;
+        case DataFormat::R8_SScaled:
+            return vk::Format::eR8Sscaled;
+        case DataFormat::R8_UInt:
+            return vk::Format::eR8Uint;
+        case DataFormat::R8_SInt:
+            return vk::Format::eR8Sint;
+        case DataFormat::R16_UNorm:
+            return vk::Format::eR16Unorm;
+        case DataFormat::R16_SNorm:
+            return vk::Format::eR16Snorm;
+        case DataFormat::R16_UScaled:
+            return vk::Format::eR16Uscaled;
+        case DataFormat::R16_SScaled:
+            return vk::Format::eR16Sscaled;
+        case DataFormat::R16_UInt:
+            return vk::Format::eR16Uint;
+        case DataFormat::R16_SInt:
+            return vk::Format::eR16Sint;
+        case DataFormat::R16_Float:
+            return vk::Format::eR16Sfloat;
+        case DataFormat::R8G8_UNorm:
+            return vk::Format::eR8G8Unorm;
+        case DataFormat::R8G8_SNorm:
+            return vk::Format::eR8G8Snorm;
+        case DataFormat::R8G8_UScaled:
+            return vk::Format::eR8G8Uscaled;
+        case DataFormat::R8G8_SScaled:
+            return vk::Format::eR8G8Sscaled;
+        case DataFormat::R8G8_UInt:
+            return vk::Format::eR8G8Uint;
+        case DataFormat::R8G8_SInt:
+            return vk::Format::eR8G8Sint;
+        case DataFormat::R8G8_SRGB:
+            return vk::Format::eR8G8Srgb;
+        case DataFormat::R8G8B8_UNorm:
+            return vk::Format::eR8G8B8Unorm;
+        case DataFormat::R8G8B8_SNorm:
+            return vk::Format::eR8G8B8Snorm;
+        case DataFormat::R8G8B8_UScaled:
+            return vk::Format::eR8G8B8Uscaled;
+        case DataFormat::R8G8B8_SScaled:
+            return vk::Format::eR8G8B8Sscaled;
+        case DataFormat::R8G8B8_UInt:
+            return vk::Format::eR8G8B8Uint;
+        case DataFormat::R8G8B8_SInt:
+            return vk::Format::eR8G8B8Sint;
+        case DataFormat::R8G8B8_SRGB:
+            return vk::Format::eR8G8B8Srgb;
+        case DataFormat::B8G8R8_UNorm:
+            return vk::Format::eB8G8R8Unorm;
+        case DataFormat::B8G8R8_SNorm:
+            return vk::Format::eB8G8R8Snorm;
+        case DataFormat::B8G8R8_UScaled:
+            return vk::Format::eB8G8R8Uscaled;
+        case DataFormat::B8G8R8_SScaled:
+            return vk::Format::eB8G8R8Sscaled;
+        case DataFormat::B8G8R8_UInt:
+            return vk::Format::eB8G8R8Uint;
+        case DataFormat::B8G8R8_SInt:
+            return vk::Format::eB8G8R8Sint;
+        case DataFormat::B8G8R8_SRGB:
+            return vk::Format::eB8G8R8Srgb;
+        case DataFormat::R32_UInt:
+            return vk::Format::eR32Uint;
+        case DataFormat::R32_SInt:
+            return vk::Format::eR32Sint;
+        case DataFormat::R32_Float:
+            return vk::Format::eR32Sfloat;
+        case DataFormat::R16G16_UNorm:
+            return vk::Format::eR16G16Unorm;
+        case DataFormat::R16G16_SNorm:
+            return vk::Format::eR16G16Snorm;
+        case DataFormat::R16G16_UScaled:
+            return vk::Format::eR16G16Uscaled;
+        case DataFormat::R16G16_SScaled:
+            return vk::Format::eR16G16Sscaled;
+        case DataFormat::R16G16_UInt:
+            return vk::Format::eR16G16Uint;
+        case DataFormat::R16G16_SInt:
+            return vk::Format::eR16G16Sint;
+        case DataFormat::R16G16_Float:
+            return vk::Format::eR16G16Sfloat;
+        case DataFormat::R8G8B8A8_UNorm:
+            return vk::Format::eR8G8B8A8Unorm;
+        case DataFormat::R8G8B8A8_SNorm:
+            return vk::Format::eR8G8B8A8Snorm;
+        case DataFormat::R8G8B8A8_UScaled:
+            return vk::Format::eR8G8B8A8Uscaled;
+        case DataFormat::R8G8B8A8_SScaled:
+            return vk::Format::eR8G8B8A8Sscaled;
+        case DataFormat::R8G8B8A8_UInt:
+            return vk::Format::eR8G8B8A8Uint;
+        case DataFormat::R8G8B8A8_SInt:
+            return vk::Format::eR8G8B8A8Sint;
+        case DataFormat::R8G8B8A8_SRGB:
+            return vk::Format::eR8G8B8A8Srgb;
+        case DataFormat::B8G8R8A8_UNorm:
+            return vk::Format::eB8G8R8A8Unorm;
+        case DataFormat::B8G8R8A8_SNorm:
+            return vk::Format::eB8G8R8A8Snorm;
+        case DataFormat::B8G8R8A8_UScaled:
+            return vk::Format::eB8G8R8A8Uscaled;
+        case DataFormat::B8G8R8A8_SScaled:
+            return vk::Format::eB8G8R8A8Sscaled;
+        case DataFormat::B8G8R8A8_UInt:
+            return vk::Format::eB8G8R8A8Uint;
+        case DataFormat::B8G8R8A8_SInt:
+            return vk::Format::eB8G8R8A8Sint;
+        case DataFormat::B8G8R8A8_SRGB:
+            return vk::Format::eB8G8R8A8Srgb;
+        case DataFormat::R16G16B16_UNorm:
+            return vk::Format::eR16G16B16Unorm;
+        case DataFormat::R16G16B16_SNorm:
+            return vk::Format::eR16G16B16Snorm;
+        case DataFormat::R16G16B16_UScaled:
+            return vk::Format::eR16G16B16Uscaled;
+        case DataFormat::R16G16B16_SScaled:
+            return vk::Format::eR16G16B16Sscaled;
+        case DataFormat::R16G16B16_UInt:
+            return vk::Format::eR16G16B16Uint;
+        case DataFormat::R16G16B16_SInt:
+            return vk::Format::eR16G16B16Sint;
+        case DataFormat::R16G16B16_Float:
+            return vk::Format::eR16G16B16Sfloat;
+        case DataFormat::R64_UInt:
+            return vk::Format::eR64Uint;
+        case DataFormat::R64_SInt:
+            return vk::Format::eR64Sint;
+        case DataFormat::R64_Float:
+            return vk::Format::eR64Sfloat;
+        case DataFormat::R32G32_UInt:
+            return vk::Format::eR32G32Uint;
+        case DataFormat::R32G32_SInt:
+            return vk::Format::eR32G32Sint;
+        case DataFormat::R32G32_Float:
+            return vk::Format::eR32G32Sfloat;
+        case DataFormat::R16G16B16A16_UNorm:
+            return vk::Format::eR16G16B16A16Unorm;;
+        case DataFormat::R16G16B16A16_SNorm:
+            return vk::Format::eR16G16B16A16Snorm;
+        case DataFormat::R16G16B16A16_UScaled:
+            return vk::Format::eR16G16B16A16Uscaled;
+        case DataFormat::R16G16B16A16_SScaled:
+            return vk::Format::eR16G16B16A16Sscaled;
+        case DataFormat::R16G16B16A16_UInt:
+            return vk::Format::eR16G16B16A16Uint;
+        case DataFormat::R16G16B16A16_SInt:
+            return vk::Format::eR16G16B16A16Sint;
+        case DataFormat::R16G16B16A16_Float:
+            return vk::Format::eR16G16B16A16Sfloat;
+        case DataFormat::R32G32B32_UInt:
+            return vk::Format::eR32G32B32Uint;
+        case DataFormat::R32G32B32_SInt:
+            return vk::Format::eR32G32B32Sint;
+        case DataFormat::R32G32B32_Float:
+            return vk::Format::eR32G32B32Sfloat;
+        case DataFormat::R64G64_UInt:
+            return vk::Format::eR64G64Uint;
+        case DataFormat::R64G64_SInt:
+            return vk::Format::eR64G64Sint;
+        case DataFormat::R64G64_Float:
+            return vk::Format::eR64G64Sfloat;
+        case DataFormat::R32G32B32A32_UInt:
+            return vk::Format::eR32G32B32A32Uint;
+        case DataFormat::R32G32B32A32_SInt:
+            return vk::Format::eR32G32B32A32Sint;
+        case DataFormat::R32G32B32A32_Float:
+            return vk::Format::eR32G32B32A32Sfloat;
+        case DataFormat::R64G64B64_UInt:
+            return vk::Format::eR64G64B64Uint;
+        case DataFormat::R64G64B64_SInt:
+            return vk::Format::eR64G64B64Sint;
+        case DataFormat::R64G64B64_Float:
+            return vk::Format::eR64G64B64Sfloat;
+        case DataFormat::R64G64B64A64_UInt:
+            return vk::Format::eR64G64B64A64Uint;
+        case DataFormat::R64G64B64A64_SInt:
+            return vk::Format::eR64G64B64A64Sint;
+        case DataFormat::R64G64B64A64_Float:
+            return vk::Format::eR64G64B64A64Sfloat;
+        case DataFormat::D16_UNorm:
+            return vk::Format::eD16Unorm;
+        case DataFormat::D32_Float:
+            return vk::Format::eD32Sfloat;
+        case DataFormat::S8_UInt:
+            return vk::Format::eS8Uint;
+        case DataFormat::D16_UNorm_S8_UInt:
+            return vk::Format::eD16UnormS8Uint;
+        case DataFormat::D24_UNorm_S8_UInt:
+            return vk::Format::eD24UnormS8Uint;
+        case DataFormat::D32_Float_S8_UInt:
+            return vk::Format::eD32SfloatS8Uint;
+        case DataFormat::Undefined:
+        default:
+            return vk::Format::eUndefined;
+        }
+    }
+
+    DataFormat DataFormatFromVkFormat(vk::Format vk_format)
+    {
+		 switch (vk_format)
+        {
+        case vk::Format::eR8Unorm:
+            return DataFormat::R8_UNorm;
+        case vk::Format::eR8Snorm:
+            return DataFormat::R8_SNorm;
+        case vk::Format::eR8Uscaled:
+            return DataFormat::R8_UScaled;
+        case vk::Format::eR8Sscaled:
+            return DataFormat::R8_SScaled;
+        case vk::Format::eR8Uint:
+            return DataFormat::R8_UInt;
+        case vk::Format::eR8Sint:
+            return DataFormat::R8_SInt;
+        case vk::Format::eR16Unorm:
+            return DataFormat::R16_UNorm;
+        case vk::Format::eR16Snorm:
+            return DataFormat::R16_SNorm;
+        case vk::Format::eR16Uscaled:
+            return DataFormat::R16_UScaled;
+        case vk::Format::eR16Sscaled:
+            return DataFormat::R16_SScaled;
+        case vk::Format::eR16Uint:
+            return DataFormat::R16_UInt;
+        case vk::Format::eR16Sint:
+            return DataFormat::R16_SInt;
+        case vk::Format::eR16Sfloat:
+            return DataFormat::R16_Float;
+        case vk::Format::eR8G8Unorm:
+            return DataFormat::R8G8_UNorm;
+        case vk::Format::eR8G8Snorm:
+            return DataFormat::R8G8_SNorm;
+        case vk::Format::eR8G8Uscaled:
+            return DataFormat::R8G8_UScaled;
+        case vk::Format::eR8G8Sscaled:
+            return DataFormat::R8G8_SScaled;
+        case vk::Format::eR8G8Uint:
+            return DataFormat::R8G8_UInt;
+        case vk::Format::eR8G8Sint:
+            return DataFormat::R8G8_SInt;
+        case vk::Format::eR8G8Srgb:
+            return DataFormat::R8G8_SRGB;
+        case vk::Format::eR8G8B8Unorm:
+            return DataFormat::R8G8B8_UNorm;
+        case vk::Format::eR8G8B8Snorm:
+            return DataFormat::R8G8B8_SNorm;
+        case vk::Format::eR8G8B8Uscaled:
+            return DataFormat::R8G8B8_UScaled;
+        case vk::Format::eR8G8B8Sscaled:
+            return DataFormat::R8G8B8_SScaled;
+        case vk::Format::eR8G8B8Uint:
+            return DataFormat::R8G8B8_UInt;
+        case vk::Format::eR8G8B8Sint:
+            return DataFormat::R8G8B8_SInt;
+        case vk::Format::eR8G8B8Srgb:
+            return DataFormat::R8G8B8_SRGB;
+        case vk::Format::eB8G8R8Unorm:
+            return DataFormat::B8G8R8_UNorm;
+        case vk::Format::eB8G8R8Snorm:
+            return DataFormat::B8G8R8_SNorm;
+        case vk::Format::eB8G8R8Uscaled:
+            return DataFormat::B8G8R8_UScaled;
+        case vk::Format::eB8G8R8Sscaled:
+            return DataFormat::B8G8R8_SScaled;
+        case vk::Format::eB8G8R8Uint:
+            return DataFormat::B8G8R8_UInt;
+        case vk::Format::eB8G8R8Sint:
+            return DataFormat::B8G8R8_SInt;
+        case vk::Format::eB8G8R8Srgb:
+            return DataFormat::B8G8R8_SRGB;
+        case vk::Format::eR32Uint:
+            return DataFormat::R32_UInt;
+        case vk::Format::eR32Sint:
+            return DataFormat::R32_SInt;
+        case vk::Format::eR32Sfloat:
+            return DataFormat::R32_Float;
+        case vk::Format::eR16G16Unorm:
+            return DataFormat::R16G16_UNorm;
+        case vk::Format::eR16G16Snorm:
+            return DataFormat::R16G16_SNorm;
+        case vk::Format::eR16G16Uscaled:
+            return DataFormat::R16G16_UScaled;
+        case vk::Format::eR16G16Sscaled:
+            return DataFormat::R16G16_SScaled;
+        case vk::Format::eR16G16Uint:
+            return DataFormat::R16G16_UInt;
+        case vk::Format::eR16G16Sint:
+            return DataFormat::R16G16_SInt;
+        case vk::Format::eR16G16Sfloat:
+            return DataFormat::R16G16_Float;
+        case vk::Format::eR8G8B8A8Unorm:
+            return DataFormat::R8G8B8A8_UNorm;
+        case vk::Format::eR8G8B8A8Snorm:
+            return DataFormat::R8G8B8A8_SNorm;
+        case vk::Format::eR8G8B8A8Uscaled:
+            return DataFormat::R8G8B8A8_UScaled;
+        case vk::Format::eR8G8B8A8Sscaled:
+            return DataFormat::R8G8B8A8_SScaled;
+        case vk::Format::eR8G8B8A8Uint:
+            return DataFormat::R8G8B8A8_UInt;
+        case vk::Format::eR8G8B8A8Sint:
+            return DataFormat::R8G8B8A8_SInt;
+        case vk::Format::eR8G8B8A8Srgb:
+            return DataFormat::R8G8B8A8_SRGB;
+        case vk::Format::eB8G8R8A8Unorm:
+            return DataFormat::B8G8R8A8_UNorm;
+        case vk::Format::eB8G8R8A8Snorm:
+            return DataFormat::B8G8R8A8_SNorm;
+        case vk::Format::eB8G8R8A8Uscaled:
+            return DataFormat::B8G8R8A8_UScaled;
+        case vk::Format::eB8G8R8A8Sscaled:
+            return DataFormat::B8G8R8A8_SScaled;
+        case vk::Format::eB8G8R8A8Uint:
+            return DataFormat::B8G8R8A8_UInt;
+        case vk::Format::eB8G8R8A8Sint:
+            return DataFormat::B8G8R8A8_SInt;
+        case vk::Format::eB8G8R8A8Srgb:
+            return DataFormat::B8G8R8A8_SRGB;
+        case vk::Format::eR16G16B16Unorm:
+            return DataFormat::R16G16B16_UNorm;
+        case vk::Format::eR16G16B16Snorm:
+            return DataFormat::R16G16B16_SNorm;
+        case vk::Format::eR16G16B16Uscaled:
+            return DataFormat::R16G16B16_UScaled;
+        case vk::Format::eR16G16B16Sscaled:
+            return DataFormat::R16G16B16_SScaled;
+        case vk::Format::eR16G16B16Uint:
+            return DataFormat::R16G16B16_UInt;
+        case vk::Format::eR16G16B16Sint:
+            return DataFormat::R16G16B16_SInt;
+        case vk::Format::eR16G16B16Sfloat:
+            return DataFormat::R16G16B16_Float;
+        case vk::Format::eR64Uint:
+            return DataFormat::R64_UInt;
+        case vk::Format::eR64Sint:
+            return DataFormat::R64_SInt;
+        case vk::Format::eR64Sfloat:
+            return DataFormat::R64_Float;
+        case vk::Format::eR32G32Uint:
+            return DataFormat::R32G32_UInt;
+        case vk::Format::eR32G32Sint:
+            return DataFormat::R32G32_SInt;
+        case vk::Format::eR32G32Sfloat:
+            return DataFormat::R32G32_Float;
+        case vk::Format::eR16G16B16A16Unorm:
+            return DataFormat::R16G16B16A16_UNorm;
+        case vk::Format::eR16G16B16A16Snorm:
+            return DataFormat::R16G16B16A16_SNorm;
+        case vk::Format::eR16G16B16A16Uscaled:
+            return DataFormat::R16G16B16A16_UScaled;
+        case vk::Format::eR16G16B16A16Sscaled:
+            return DataFormat::R16G16B16A16_SScaled;
+        case vk::Format::eR16G16B16A16Uint:
+            return DataFormat::R16G16B16A16_UInt;
+        case vk::Format::eR16G16B16A16Sint:
+            return DataFormat::R16G16B16A16_SInt;
+        case vk::Format::eR16G16B16A16Sfloat:
+            return DataFormat::R16G16B16A16_Float;
+        case vk::Format::eR32G32B32Uint:
+            return DataFormat::R32G32B32_UInt;
+        case vk::Format::eR32G32B32Sint:
+            return DataFormat::R32G32B32_SInt;
+        case vk::Format::eR32G32B32Sfloat:
+            return DataFormat::R32G32B32_Float;
+        case vk::Format::eR64G64Uint:
+            return DataFormat::R64G64_UInt;
+        case vk::Format::eR64G64Sint:
+            return DataFormat::R64G64_SInt;
+        case vk::Format::eR64G64Sfloat:
+            return DataFormat::R64G64_Float;
+        case vk::Format::eR32G32B32A32Uint:
+            return DataFormat::R32G32B32A32_UInt;
+        case vk::Format::eR32G32B32A32Sint:
+            return DataFormat::R32G32B32A32_SInt;
+        case vk::Format::eR32G32B32A32Sfloat:
+            return DataFormat::R32G32B32A32_Float;
+        case vk::Format::eR64G64B64Uint:
+            return DataFormat::R64G64B64_UInt;
+        case vk::Format::eR64G64B64Sint:
+            return DataFormat::R64G64B64_SInt;
+        case vk::Format::eR64G64B64Sfloat:
+            return DataFormat::R64G64B64_Float;
+        case vk::Format::eR64G64B64A64Uint:
+            return DataFormat::R64G64B64A64_UInt;
+        case vk::Format::eR64G64B64A64Sint:
+            return DataFormat::R64G64B64A64_SInt;
+        case vk::Format::eR64G64B64A64Sfloat:
+            return DataFormat::R64G64B64A64_Float;
+        case vk::Format::eD16Unorm:
+            return DataFormat::D16_UNorm;
+        case vk::Format::eD32Sfloat:
+            return DataFormat::D32_Float;
+        case vk::Format::eS8Uint:
+            return DataFormat::S8_UInt;
+        case vk::Format::eD16UnormS8Uint:
+            return DataFormat::D16_UNorm_S8_UInt;
+        case vk::Format::eD24UnormS8Uint:
+            return DataFormat::D24_UNorm_S8_UInt;
+        case vk::Format::eD32SfloatS8Uint:
+            return DataFormat::D32_Float_S8_UInt;
+        default:
+            return DataFormat::Undefined;
+        }
+    }
 }
