+++ BRenderer/Renderer/Vulkan/VulkanInc.h
@@ -7,6 +7,7 @@
 #define VULKAN_HPP_NO_EXCEPTIONS
 #define VULKAN_HPP_DISPATCH_LOADER_DYNAMIC 1
 #include "vulkan/vulkan.hpp"
+#include "vulkan/vulkan_hash.hpp"
 
 #include "vk_mem_alloc.h"
