+++ BRenderer/Visualization/WindowRenderer.cpp
@@ -3,8 +3,6 @@
 #include <Visualization/SceneRenderer.h>
 #include <Visualization/Window.h>
 #include <Renderer/Vulkan/VulkanRenderDevice.h>
-#include <Renderer/Swapchain.h>
-#include <Renderer/Shader.h>
 #include <Core/LogSystem.h>
